backend-fastapi/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.9
motor==3.7.1
pydantic==2.11.4
pydantic-settings==2.9.1
python-docx==1.1.2
PyMuPDF==1.25.5
openai==1.76.0
groq==0.10.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.115.0
apscheduler==3.10.4
authlib==1.3.1
python-jose[cryptography]==3.3.0
cryptography==42.0.7
PyJWT==2.12.1
email-validator==2.2.0
